feat(core): extract plugins/ into core/ kernel (Phase E)
Move all 12 Fastify plugins and colocated tests to
backend/src/core/plugins/. Delete dead request-context.ts (zero imports,
functionality merged into request-tracing.ts). Temporarily adjust
service imports in auth.ts, request-tracing.ts, and socket-io.ts to
resolve through ../../services/ until Phase F/G moves those to core/.
Create re-export shims at original paths.

Part of #706
