+++ backend/src/plugins/dev-origins.ts
@@ -1,8 +1,2 @@
-/**
- * Single source of truth for development-mode allowed origins.
- * Used by both the CORS plugin and Socket.IO plugin to avoid drift.
- */
-export const DEV_ALLOWED_ORIGINS = [
-  'http://localhost:5273',
-  'http://localhost:8080',
-];
+// Shim — re-exports from core/plugins (will be removed in Phase H)
+export { DEV_ALLOWED_ORIGINS } from '../core/plugins/dev-origins.js';
